// apps/legal-gazette-api/src/core/enums/status.enum.ts
export enum StatusIdEnum {
  SUBMITTED = 'cd3bf301-52a1-493e-8c80-a391c310c840',
  IN_PROGRESS = '7ef679c4-4f66-4892-b6ad-ee05e0be4359',
  READY_FOR_PUBLICATION = 'a2f3b1c4-2d5e-4a7b-8c6f-9d1e0f3a2b8c',
  IN_PUBLISHING = '5d1561dc-693c-4682-9b70-03b664d65d9b',
  PUBLISHED = 'bd835a1d-0ecb-4aa4-9910-b5e60c30dced',
  REJECTED = 'f3a0b1c4-2d5e-4a7b-8c6f-9d1e0f3a2b8c',
  WITHDRAWN = 'e2f3b1c4-2d5e-4a7b-8c6f-9d1e0f3a2b8c',
}
export enum StatusEnum {
  SUBMITTED = 'Innsent',
  IN_PROGRESS = 'Í vinnslu',
  IN_PUBLISHING = 'Í útgáfu',
  READY_FOR_PUBLICATION = 'Tilbúið til útgáfu',
  PUBLISHED = 'Útgefið',
  REJECTED = 'Hafnað',
  REVOKED = 'Afturkallað',
}